--- Assets/Scripts/Monsters/Scholar/Scholar.Appearance.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ea5c51f8384544747b79386ddf6b3788
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
