--- Assets/Scripts/Monsters/Scholar/Scholar.BeHit.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43b31033b30fd2c49a9f44c6af9cf043
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
