+++ Assets/Scripts/Monsters/Scholar/Scholar.cs
@@ -6,11 +6,26 @@ using UnityEngine.Events;
 using TMPro;
 using System;
 
-public partial class Scholar : LivingEntity
+public class Scholar : LivingEntity
 {
+    private SpriteRenderer scholarSpriteRenderer;
+    private TMP_Text hpTextBox;
+    private Color BehitColor;
+    private Color fadeColor;
+    
+    private float transparent = 0.5f;
+    private float normal = 1.0f;
+    
     private ScholarStateMachine scholarStateMachine;
-
     public ScholarManager scholarManager;
+    
+    public bool isIdle = false;
+
+    public bool IsIdle
+    {
+        get { return isIdle; }
+        set { isIdle = value; }
+    }
 
     private void Awake()
     {
@@ -28,11 +43,6 @@ public partial class Scholar : LivingEntity
         this.scholarManager = transform.parent.GetComponent<ScholarManager>();
     }
 
-    private GameObject GetChild(int v)
-    {
-        throw new NotImplementedException();
-    }
-
     void Start()
     {
         this.scholarStateMachine.Initialize("Appearance", this);
@@ -40,9 +50,55 @@ public partial class Scholar : LivingEntity
         // TO DO: 20240225 kimyeonmo UI 로직 분리 (delegate 이용)
         this.hpTextBox.text = HP.ToString();
     }
-
-    void Update()
+    
+    private void OnTriggerEnter2D(Collider2D collision)
     {
+        if (collision.gameObject.tag == "Bullet")
+        {
+            if(this.IsIdle == true)
+            {
+                Debug.Log("총알 맞았다!!");
 
+                scholarManager.SetSchloarBehit(true);
+                OnDamage(523456789);
+
+                Debug.Log("비히트?? " + scholarManager.GetIsSchloarBehit());
+
+                this.hpTextBox.text = HP.ToString();
+
+                StartCoroutine(BeHitEffect());
+            }
+        }
+    }
+    
+    public IEnumerator AppearanceCoroutine()
+    {
+        this.fadeColor = scholarSpriteRenderer.color;
+        this.fadeColor.a = 0f;
+        scholarSpriteRenderer.color = this.fadeColor;
+
+        while (this.fadeColor.a <= 1f)
+        {
+            this.fadeColor = scholarSpriteRenderer.color;
+            this.fadeColor.a += 0.05f;
+            scholarSpriteRenderer.color = this.fadeColor;
+            
+            yield return new WaitForSeconds(0.05f);
+        }
+    }
+
+    private IEnumerator BeHitEffect()
+    {
+        this.BehitColor = this.scholarSpriteRenderer.color;
+
+        this.BehitColor.a = this.transparent;
+
+        this.scholarSpriteRenderer.color = this.BehitColor;
+
+        yield return new WaitForSeconds(0.3f);
+
+        this.BehitColor.a = this.normal;
+
+        this.scholarSpriteRenderer.color = this.BehitColor;
     }
 }
