--- Assets/Scripts/Monsters/Scholar/ScholarUIController.cs
@@ -1,16 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class ScholarUIController : MonoBehaviour
-{
-    void Start()
-    {
-        
-    }
-
-    void Update()
-    {
-        
-    }
-}
