--- Assets/Scripts/Monsters/Scholar/ScholarUIController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d2366dcdbef26f44f9bcd46b1e6cc0eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
